use np in protocol interfaces

+++ src/satsim/protocols.py
@@ -1,5 +1,7 @@
 from typing import Protocol, runtime_checkable
 
+import numpy as np
+
 from satsim.environment.earth import Environment
 from satsim.spacecraft import Spacecraft
 from satsim.types import StateVector, Vec3
@@ -7,16 +9,16 @@ from satsim.types import StateVector, Vec3
 
 @runtime_checkable
 class ForceModel(Protocol):
-    def acceleration(self, state: StateVector, t: float, env: Environment) -> Vec3: ...
+    def acceleration(self, state: StateVector, t: np.float64, env: Environment) -> Vec3: ...
 
 
 @runtime_checkable
 class StopCondition(Protocol):
-    def check(self, state: StateVector, t: float) -> bool: ...
+    def check(self, state: StateVector, t: np.float64) -> bool: ...
 
 
 @runtime_checkable
 class Event(Protocol):
-    trigger_time: float
+    trigger_time: np.float64
 
     def trigger(self, spacecraft: Spacecraft, env: Environment) -> None: ...
